requirements.txt
pymysql
pandas
python-dotenv
requests